Moved constants to separate folder in helper files
