+++ collect_spreadsheet_metadata.py
@@ -1,6 +1,4 @@
 import os
-import re
-import shutil
 import argparse
 import pandas as pd
