+++ merge_file_manifest.py
@@ -1,11 +1,8 @@
 import os
-import re
 import argparse
-from collections import defaultdict
-
 import pandas as pd
 
-from helper_files.file_mapping import FILE_MANIFEST_MAPPING, TIER_1_MAPPING, FASTQ_STANDARD_FIELDS
+from helper_files.constants.file_mapping import FILE_MANIFEST_MAPPING, TIER_1_MAPPING, FASTQ_STANDARD_FIELDS
 
 LAST_BIOMATERIAL = 'cell_suspension.biomaterial_core.biomaterial_id'
