+++ merge_tier2_metadata.py
@@ -5,8 +5,8 @@ from collections import defaultdict
 
 import pandas as pd
 
-from helper_files.tier2_mapping import TIER2_TO_DCP, TIER2_TO_DCP_UPDATE, LUNG_DIGESTION, TIER2_MANUAL_FIX
-from helper_files.dcp_required import dcp_required_entities, dcp_required_modules
+from helper_files.constants.tier2_mapping import TIER2_TO_DCP, TIER2_TO_DCP_UPDATE, LUNG_DIGESTION, TIER2_MANUAL_FIX
+from helper_files.constants.dcp_required import dcp_required_entities, dcp_required_modules
 from convert_to_dcp import fill_missing_ontology_ids, fill_ontology_labels
 
 # Small helpers
